requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib[bcrypt]==1.7.4
python-jose[cryptography]
python-multipart
pydantic
bcrypt==4.0.1

